+++ bootstrap-conduits.sh
@@ -63,7 +63,7 @@ script_path() {
 # CONFIG
 ########################################
 IMAGE="ghcr.io/psiphon-inc/conduit/cli:latest"
-STACK_VERSION="2026.03.01.32"
+STACK_VERSION="2026.03.01.33"
 BASE_PORT=9090
 GRAFANA_PORT=3000
 BACKUP_DIR="./backups"